+++ src/textual/layouts/vertical.py
@@ -3,43 +3,55 @@ from __future__ import annotations
 from typing import Iterable
 
 from rich.console import Console
+from rich.measure import Measurement
 
 from .. import log
 from ..geometry import Offset, Region, Size
-from ..layout import Layout
+from ..layout import Layout, WidgetPlacement
 from ..layout_map import LayoutMap
 from ..widget import Widget
 
 
 class VerticalLayout(Layout):
-    def __init__(self, *, z: int = 0, gutter: tuple[int, int] | None = None):
+    def __init__(
+        self,
+        *,
+        auto_width: bool = False,
+        z: int = 0,
+        gutter: tuple[int, int] | None = None
+    ):
+        self.auto_width = auto_width
         self.z = z
         self.gutter = gutter or (0, 0)
         self._widgets: list[Widget] = []
+        self._max_widget_width = 0
         super().__init__()
 
     def add(self, widget: Widget) -> None:
         self._widgets.append(widget)
+        self._max_widget_width = max(widget.app.measure(widget), self._max_widget_width)
 
     def clear(self) -> None:
         del self._widgets[:]
+        self._max_widget_width = 0
 
     def get_widgets(self) -> Iterable[Widget]:
         return self._widgets
 
-    def generate_map(
-        self, console: Console, size: Size, viewport: Region, scroll: Offset
-    ) -> LayoutMap:
+    def arrange(self, size: Size, scroll: Offset) -> Iterable[WidgetPlacement]:
         index = 0
-        width, height = size
+        width, _height = size
         gutter_height, gutter_width = self.gutter
-        render_width = width - gutter_width * 2
+        render_width = (
+            max(width, self._max_widget_width) + gutter_width * 2
+            if self.auto_width
+            else width - gutter_width * 2
+        )
+
         x = gutter_width
         y = gutter_height
-        map: LayoutMap = LayoutMap(size)
 
-        def add_widget(widget: Widget, region: Region, clip: Region) -> None:
-            map.add_widget(console, widget, region, (self.z, index), clip)
+        total_region = size.region
 
         for widget in self._widgets:
             if (
@@ -50,6 +62,7 @@ class VerticalLayout(Layout):
             assert widget.render_cache is not None
             render_height = widget.render_cache.size.height
             region = Region(x, y, render_width, render_height)
-            add_widget(widget, region - scroll, viewport)
+            yield WidgetPlacement(region, widget, (self.z, index))
+            total_region = total_region.union(region)
 
-        return map
+        yield WidgetPlacement(total_region.expand(self.gutter))