+++ src/textual/widgets/_tree_control.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
-from functools import lru_cache
 
 from typing import Generic, Iterator, NewType, TypeVar
 
+import rich.repr
 from rich.console import RenderableType
 from rich.text import Text, TextType
 from rich.tree import Tree
@@ -15,7 +15,7 @@ from ..reactive import Reactive
 from .._types import MessageTarget
 from ..widget import Widget
 from ..message import Message
-from ..messages import CursorMoveMessage
+from ..messages import CursorMove
 
 
 NodeID = NewType("NodeID", int)
@@ -24,6 +24,7 @@ NodeID = NewType("NodeID", int)
 NodeDataType = TypeVar("NodeDataType")
 
 
+@rich.repr.auto
 class TreeNode(Generic[NodeDataType]):
     def __init__(
         self,
@@ -45,6 +46,11 @@ class TreeNode(Generic[NodeDataType]):
         self._empty = False
         self._tree.expanded = False
         self.children: list[TreeNode] = []
+
+    def __rich_repr__(self) -> rich.repr.Result:
+        yield "id", self.id
+        yield "label", self.label
+        yield "data", self.data
 
     @property
     def control(self) -> TreeControl:
@@ -154,10 +160,14 @@ class TreeNode(Generic[NodeDataType]):
         return self._control.render_node(self)
 
 
+@rich.repr.auto
 class TreeClick(Generic[NodeDataType], Message, bubble=True):
     def __init__(self, sender: MessageTarget, node: TreeNode[NodeDataType]) -> None:
         self.node = node
         super().__init__(sender)
+
+    def __rich_repr__(self) -> rich.repr.Result:
+        yield "node", self.node
 
 
 class TreeControl(Generic[NodeDataType], Widget):
@@ -189,11 +199,11 @@ class TreeControl(Generic[NodeDataType], Widget):
     show_cursor: Reactive[bool] = Reactive(False, layout=True)
 
     def watch_show_cursor(self, value: bool) -> None:
-        self.emit_no_wait(CursorMoveMessage(self, self.cursor_line))
+        self.emit_no_wait(CursorMove(self, self.cursor_line))
 
     def watch_cursor_line(self, value: int) -> None:
         if self.show_cursor:
-            self.emit_no_wait(CursorMoveMessage(self, value + self.gutter.top))
+            self.emit_no_wait(CursorMove(self, value + self.gutter.top))
 
     async def add(
         self,
@@ -309,7 +319,7 @@ if __name__ == "__main__":
         async def on_mount(self, event: events.Mount) -> None:
             await self.view.dock(TreeControl("Tree Root", data="foo"))
 
-        async def message_tree_click(self, message: TreeClick) -> None:
+        async def handle_tree_click(self, message: TreeClick) -> None:
             if message.node.empty:
                 await message.node.add("foo")
                 await message.node.add("bar")