+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.11] - Unreleased
+
+### Changed
+
+- Changed message handlers to use prefix handle\_
+- Renamed messages to drop the Message suffix
+- Events now bubble by default
+- Refactor of layout
+
+### Added
+
+- Added App.measure
+- Added auto_width to Vertical Layout, WindowView, an ScrollView
+- Added big_table.py example
+- Added easing.py example
+
 ## [0.1.10] - 2021-08-25
 
 ### Added